+++ addOns/customactivescan/src/main/java/org/zaproxy/zap/extension/customactivescan/view/MainWorkPanelTab.java
@@ -16,7 +16,7 @@ public class MainWorkPanelTab extends AbstractPanel {
         setLayout(new CardLayout());
         this.setName("CustomActiveScan");
         this.setIcon(ExtensionAscanRules.cIcon);
-        CustomScanMainPanel mainPanel = new CustomScanMainPanel();
+        CustomScanMainPanel mainPanel = new CustomScanMainPanel(this);
         extensionAscan.setCustomScanMainPanel(mainPanel);
         Border mainBorder = new LineBorder(Color.RED, 1);
         mainPanel.setBorder(mainBorder);